fix(transaction-query): move util function to api/service package
